Changed forts
Forts are now fully able to be part of armies. Some methods from unit were moved to BattlefieldEntity.
Most of these methods return simply nothing or nothing of value, because they will only be used on units.

But this means forts are able to be part of armies, and attacked.